backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.19
psycopg2-binary>=2.9.7
alembic>=1.11.1
pydantic>=2.1.1
pydantic-settings>=2.0.2
strawberry-graphql[fastapi]>=0.201.1
websockets>=11.0.3
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
redis>=4.6.0
pytest>=7.4.0
pytest-asyncio>=0.21.1
httpx>=0.24.1
